Modification du model pour creer en masse samples et containers à partir fichier CSV
git-svn-id: https://gsphere.genoscope.cns.fr/svn/NGL/trunk/ngl@541 d0e871a8-0b98-4e59-9d77-5ea3c331444f

+++ lib-ngl-common/app/models/laboratory/container/instance/Volume.java
@@ -12,7 +12,7 @@ import java.util.Date;
  */
 public class Volume {
 
-	public Float value;
+	public Double value;
 	public Date date;
 	public String unit;
